Add files via upload
Was able to slow down the web page crawling speed. 
Wasn't able to get the data from the graph into a table (it was a aws link and not public access)
I found another chart with more data here https://data.humdata.org/visualization/ukraine-humanitarian-operations/
and may be able to scrape it, stay tuned.